:white_check_mark: add・update candle tests
