+++ quant_tick/tests/models/test_candles.py
@@ -1,23 +1,19 @@
+from datetime import datetime
+
 import pandas as pd
+import time_machine
 from django.test import TestCase
 
-from quant_tick.constants import Exchange, FileData, Frequency, SampleType
+from quant_tick.constants import FileData, Frequency, SampleType
 from quant_tick.lib import get_current_time, get_min_time, get_previous_time
-from quant_tick.models import Candle, CandleCache, Symbol, TradeData
+from quant_tick.models import Candle, CandleCache, TradeData
 from quant_tick.storage import convert_candle_cache_to_daily
 
 from ..base import BaseSymbolTest, BaseWriteTradeDataTest
+from .candle_types.base import BaseDayIteratorTest
 
 
-class BaseCandleTest(TestCase):
-    def get_symbol(self, name: str, exchange: Exchange = Exchange.COINBASE) -> Symbol:
-        return Symbol.objects.create(
-            exchange=exchange,
-            api_symbol=name,
-        )
-
-
-class CandleTest(BaseWriteTradeDataTest, BaseCandleTest):
+class CandleDataFrameTest(BaseWriteTradeDataTest, TestCase):
     def setUp(self):
         super().setUp()
         self.timestamp_to = self.timestamp_from + pd.Timedelta("1min")
@@ -44,12 +40,111 @@ class CandleTest(BaseWriteTradeDataTest, BaseCandleTest):
         self.assertTrue(all(data_frame == df))
 
 
-class CandleCacheTest(BaseSymbolTest, BaseCandleTest):
+@time_machine.travel(datetime(2009, 1, 4), tick=False)
+class CandleInitializeTest(BaseSymbolTest, BaseDayIteratorTest, TestCase):
     def setUp(self):
         super().setUp()
-        self.symbol = self.get_symbol("test")
         self.candle = Candle.objects.create(
-            symbol=self.symbol, json_data={"sample_type": SampleType.NOTIONAL}
+            symbol=self.get_symbol(), json_data={"source_data": FileData.RAW}
+        )
+
+    def create_candle_cache(self, timestamp: datetime) -> CandleCache:
+        return CandleCache.objects.create(
+            candle=self.candle, timestamp=timestamp, frequency=Frequency.DAY
+        )
+
+    def test_initial_timestamp_from_without_candle_date_from(self):
+        timestamp_from, timestamp_to, _ = self.candle.initialize(
+            self.timestamp_from, self.three_days_from_now
+        )
+        self.assertEqual(timestamp_from, self.timestamp_from)
+        self.assertEqual(timestamp_to, self.three_days_from_now)
+
+    def test_initial_timestamp_from_with_candle_date_from(self):
+        self.candle.date_from = self.one_day_from_now.date()
+        timestamp_from, timestamp_to, _ = self.candle.initialize(
+            self.timestamp_from, self.three_days_from_now
+        )
+        self.assertEqual(timestamp_from, self.one_day_from_now)
+
+    def test_initial_timestamp_from_with_symbol_date_from(self):
+        symbol = self.candle.symbol
+        symbol.date_from = self.one_day_from_now.date()
+        symbol.save()
+        timestamp_from, timestamp_to, _ = self.candle.initialize(
+            self.timestamp_from, self.three_days_from_now
+        )
+        self.assertEqual(timestamp_from, self.one_day_from_now)
+        self.assertEqual(timestamp_to, self.three_days_from_now)
+
+    def test_initial_timestamp_from_uses_later_symbol_or_candle_date_from(self):
+        symbol = self.candle.symbol
+        symbol.date_from = self.two_days_from_now.date()
+        symbol.save()
+        self.candle.date_from = self.one_day_from_now.date()
+        timestamp_from, timestamp_to, _ = self.candle.initialize(
+            self.timestamp_from, self.three_days_from_now
+        )
+        self.assertEqual(timestamp_from, self.two_days_from_now)
+        self.assertEqual(timestamp_to, self.three_days_from_now)
+
+    def test_initial_timestamp_from_skips_before_symbol_date_from(self):
+        symbol = self.candle.symbol
+        symbol.date_from = self.three_days_from_now.date()
+        symbol.save()
+        timestamp_from, timestamp_to, data = self.candle.initialize(
+            self.timestamp_from, self.two_days_from_now
+        )
+        self.assertEqual(timestamp_from, self.two_days_from_now)
+        self.assertEqual(timestamp_to, self.two_days_from_now)
+        self.assertEqual(data, {})
+
+    def test_initial_timestamp_from_with_candle_cache(self):
+        for i in range(2):
+            self.create_candle_cache(self.timestamp_from + pd.Timedelta(f"{i}d"))
+        timestamp_from, timestamp_to, _ = self.candle.initialize(
+            self.timestamp_from, self.three_days_from_now
+        )
+        self.assertEqual(timestamp_from, self.two_days_from_now)
+        self.assertEqual(timestamp_to, self.three_days_from_now)
+
+    def test_initial_timestamp_from_with_candle_cache_and_retry(self):
+        for i in range(2):
+            self.create_candle_cache(self.timestamp_from + pd.Timedelta(f"{i}d"))
+        timestamp_from, timestamp_to, _ = self.candle.initialize(
+            self.timestamp_from, self.three_days_from_now, retry=True
+        )
+        self.assertEqual(timestamp_from, self.timestamp_from)
+        self.assertEqual(timestamp_to, self.three_days_from_now)
+
+    def test_initial_timestamp_from_with_both_candle_date_from_and_candle_cache(self):
+        self.candle.date_from = self.one_day_from_now.date()
+        for i in range(2):
+            self.create_candle_cache(self.timestamp_from + pd.Timedelta(f"{i}d"))
+        timestamp_from, timestamp_to, _ = self.candle.initialize(
+            self.timestamp_from, self.three_days_from_now
+        )
+        self.assertEqual(timestamp_from, self.two_days_from_now)
+        self.assertEqual(timestamp_to, self.three_days_from_now)
+
+    def test_initial_timestamp_from_with_both_candle_date_from_candle_cache_and_retry(
+        self,
+    ):
+        self.candle.date_from = self.one_day_from_now.date()
+        for i in range(3):
+            self.create_candle_cache(self.timestamp_from + pd.Timedelta(f"{i}d"))
+        timestamp_from, timestamp_to, _ = self.candle.initialize(
+            self.timestamp_from, self.three_days_from_now, retry=True
+        )
+        self.assertEqual(timestamp_from, self.one_day_from_now)
+        self.assertEqual(timestamp_to, self.three_days_from_now)
+
+
+class CandleCacheTest(BaseSymbolTest, TestCase):
+    def setUp(self):
+        super().setUp()
+        self.candle = Candle.objects.create(
+            symbol=self.get_symbol(), json_data={"sample_type": SampleType.NOTIONAL}
         )
 
     def test_convert_candle_cache_to_daily(self):
